kubernetes/requirements.txt
kopf==0.25
kubernetes==10.0.1
python-consul==1.1.0
PyYAML==5.3
